5.x-feat-spid-test-refactoring
