--- src/test/java/it/smartcommunitylab/aac/spid/SpidSecurityTest.java
@@ -1,657 +1,0 @@
-package it.smartcommunitylab.aac.spid;
-
-import com.github.tomakehurst.wiremock.WireMockServer;
-import com.maciejwalkowiak.wiremock.spring.ConfigureWireMock;
-import com.maciejwalkowiak.wiremock.spring.EnableWireMock;
-import com.maciejwalkowiak.wiremock.spring.InjectWireMock;
-import it.smartcommunitylab.aac.identity.model.ConfigurableIdentityProvider;
-import it.smartcommunitylab.aac.spid.provider.IdentityProvider;
-import it.smartcommunitylab.aac.spid.setup.BaseSpidTest;
-import it.smartcommunitylab.aac.spid.setup.MockIdpSpid;
-import it.smartcommunitylab.aac.spid.setupflow.SpidRequest;
-import it.smartcommunitylab.aac.spid.utils.HackerUtils;
-import it.smartcommunitylab.aac.spid.setupflow.SpidRequestFlow;
-import it.smartcommunitylab.aac.spid.setupflow.SpidResponseBuilder;
-import it.smartcommunitylab.aac.spid.utils.UserUtils;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpSession;
-import org.springframework.security.web.WebAttributes;
-import org.springframework.test.context.ActiveProfiles;
-
-import java.util.List;
-
-import javax.transaction.Transactional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-/**
- * Test suite for validating the security and resilience of the SPID SAML 2.0 authentication flow against malicious attacks.
- * Verifies robust mitigation against replay attacks, CSRF (RelayState manipulation), open redirects, signature stripping, and assertion tampering.
- */
-@SpringBootTest
-@AutoConfigureMockMvc
-@ActiveProfiles({"test", "test-spid"})
-@EnableWireMock({
-    // Setup two fixed-port WireMock servers, mapping them to their respective YAML configuration properties
-    @ConfigureWireMock(port = 58838, name = "idp-server-redirect", property = "wiremock.idp.redirect.url"),
-    @ConfigureWireMock(port = 58839, name = "idp-server-post", property = "wiremock.idp.post.url")
-})
-// Add @Transactional to clean up the DB automatically between @Test methods within this class
-@Transactional
-public class SpidSecurityTest extends BaseSpidTest {
-
-    // Inject Redirect WireMock
-    @InjectWireMock("idp-server-redirect")
-    protected WireMockServer mockIdPServerRedirect;
-
-    // Inject Post WireMock
-    @InjectWireMock("idp-server-post")
-    protected WireMockServer mockIdPServerPost;
-
-    protected UserUtils userUtils = new UserUtils();
-    protected HackerUtils hackerUtils = new HackerUtils();
-    protected MockIdpSpid mockIdpSpid = new MockIdpSpid();
-    protected IdentityProvider identityProvider = new IdentityProvider();
-
-    @BeforeEach
-    public void setupConfigurationAndMocks() {
-        initMockMvc();
-        mockIdpSpid.prepareMockMetadata(mockIdPServerRedirect, mockIdPServerPost);
-
-        config.getRealms().forEach(realm -> {
-            if ("spid-test".equals(realm.getRealm().getSlug())) {
-                List<ConfigurableIdentityProvider> idps = realm.getIdentityProviders();
-
-                // Any Identity Provider loaded from the bootstrap can be used here
-                ConfigurableIdentityProvider idp = idps.get(3);
-
-                identityProvider.initRealmByBoostrap(idp, BASE_URL, METADATA_PATH, SSO_PATH);
-                identityProvider.initRegistrationIdBinding(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_POST);
-            }
-        });
-    }
-
-    @Test
-    @DisplayName("Autenticazione con successo: HTTP-Redirect Binding")
-    public void testAuthenticationSucceedsWithRedirectBinding() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect) // REDIRECT
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, spidRequest.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, identityProvider.signingIdpEntityId) // REDIRECT
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .withSignature()
-            .buildResponse();
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            )
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(USER_DESTINATION_URL));
-    }
-
-    @Test
-    @DisplayName("Autenticazione con successo: HTTP-POST Binding")
-    public void testAuthenticationSucceedsWithPostBinding() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdPost) // POST
-            .withPostBinding(true)
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, spidRequest.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_POST, identityProvider.signingIdpEntityId) // POST
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .withSignature()
-            .buildResponse();
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            )
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(USER_DESTINATION_URL));
-    }
-
-    /**
-     * Security Test: Ensures authentication fails and the payload is rejected
-     * if the SAML Response lacks a valid cryptographic signature.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso se la SAML Response NON è firmata")
-    public void testAuthenticationFailsOnUnsignedSamlResponse() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, spidRequest.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, identityProvider.signingIdpEntityId)
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .buildResponse();
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response) // UNSIGNED Payload (Rejected by SP)
-                .param("RelayState", spidRequest.getRelayState())     // Valid State
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            )
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL))
-            .andReturn();
-
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-        assertThat(sessionException.getMessage()).contains("1000"); // Cryptographic validation error code
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the RelayState parameter
-     * is missing, unexpected, or has been manipulated during the flow.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso se il RelayState è inesistente o manipolato")
-    public void testAuthenticationFailsOnMissingOrManipulatedRelayState() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, spidRequest.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, identityProvider.signingIdpEntityId)
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .withSignature()
-            .buildResponse();
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", "SSdfMklR7My_NohjAY72i57SfjumEOEhIXnEYVVMIew=") // MANIPULATED RelayState (Invalid CSRF)
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            )
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL))
-            .andReturn();
-
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-        assertThat(sessionException.getMessage()).contains("No relying party registration found");
-    }
-
-    /**
-     * Security Test: Verifies protection against Replay Attacks.
-     * The system must reject a SAML Response that has already been processed.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Attacco di Replay (SAML Response riutilizzata da un utente terzo)")
-    public void testAuthenticationFailsOnReplayAttack() throws Exception {
-        SpidRequest victimSpidResponse = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, victimSpidResponse.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, identityProvider.signingIdpEntityId)
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .withSignature()
-            .buildResponse();
-
-        // FIRST USE (Victim completes login successfully)
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", victimSpidResponse.getRelayState())
-                .session(victimSpidResponse.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(USER_DESTINATION_URL));
-
-        // ATTACK: attacker intercepts the signed SAMLResponse and attempts to replay it.
-        MockHttpSession hackerSession = userUtils.createSessionWithSavedClientRequest(BASE_URL);
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response) // REPLAYED Payload (Intercepted from victim)
-                .param("RelayState", victimSpidResponse.getRelayState())     // REPLAYED RelayState
-                .session(hackerSession) // DIFFERENT Session (The Hacker's session)
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andReturn();
-
-        Exception hackerException = (Exception) hackerSession.getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(hackerException).isNotNull();
-        assertThat(hackerException.getMessage()).contains("No relying party registration found");
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the 'InResponseTo' attribute
-     * in the SAML Response does not match the ID of the original AuthnRequest.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per InResponseTo errato (Risposta non corrispondente alla richiesta)")
-    public void testAuthenticationFailsOnInvalidInResponseTo() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // InResponseTo errato (Mismatch Request ID)
-        String response = hackerUtils.prepareForSimulationNotValidRequestId(
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response) // MALICIOUS Payload (Signed, but mismatched InResponseTo)
-                .param("RelayState", spidRequest.getRelayState())   // Valid State
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the SAML Response is intended
-     * for a different Service Provider (Audience Restriction validation).
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Audience errata (Risposta destinata a un altro Service Provider)")
-    public void testAuthenticationFailsOnInvalidAudience() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Audience errata (Invalid EntityID)
-        String response = hackerUtils.prepareForSimulationNotValidEntityId(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response) // MALICIOUS Payload (Signed, but wrong Audience)
-                .param("RelayState", spidRequest.getRelayState())   // Valid State
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the Identity Provider
-     * returns a SPID security level lower than the one requested.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso se il livello SPID restituito (L1) è inferiore a quello richiesto (L2)")
-    public void testAuthenticationFailsOnLowerSpidLevel() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Downgrade Livello SPID (L2 -> L1)
-        String response = hackerUtils.prepareForSimulationNotValidChangeSpidLevelLow(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        // Verify error is 1000 (errore di validazione del payload)
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-        assertThat(sessionException.getMessage()).contains("1000");
-    }
-
-    /**
-     * Security Test: Ensures authentication succeed if the Identity Provider
-     * returns a SPID security level higher than the one requested.
-     */
-    @Test
-    @DisplayName("Autenticazione con successo: livello SPID restituito (L3) è superiore a quello richiesto (L2)")
-    public void testAuthenticationSucceedOnHighSpidLevel() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Downgrade SPID level (L2 -> L3)
-        String response = hackerUtils.prepareForSimulationNotValidChangeSpidLevelHigh(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(USER_DESTINATION_URL));
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the SAML Assertion
-     * has expired (NotOnOrAfter condition is in the past).
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Assertion scaduta")
-    public void testAuthenticationFailsOnExpiredAssertion() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Expired Assertion (NotOnOrAfter in the past)
-        String response = hackerUtils.prepareForSimulationNotValidNotOnOrAfter(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        // Verify that Spring Security rejected the Assertion due to time restrictions
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-        assertThat(sessionException.getMessage()).containsIgnoringCase("is no longer valid");
-    }
-
-    /**
-     * Security Concept: Open Redirect Vulnerability Mitigation.
-     * * In a classic Open Redirect attack against SAML SSO, an attacker crafts a malicious link
-     * and attempts to inject an external URL (e.g., "https://phishing-site.com") into the RelayState
-     * or the initial login request. If the Service Provider blindly trusts this input post-authentication,
-     * the user is automatically redirected to the malicious site after a successful SPID login.
-     * 1. It uses an opaque, securely generated identifier for the RelayState (not a raw URL).
-     * 2. The actual intended destination URL is safely stored in the HttpSession (via HttpSessionRequestCache)
-     * and never transmitted in clear text to the IdP.
-     * 3. Any attempt to pass a raw URL as the RelayState during the SAML Response phase will fail
-     * the cryptographic context lookup, resulting in a safe authentication failure instead of a redirect.
-     */
-    @Test
-    @DisplayName("Sicurezza: Mitigazione nativa contro attacchi di Open Redirect")
-    public void testAuthenticationFailsOnOpenRedirectAttempt() throws Exception {
-        // 1. Attacker crafts a malicious URL intended for an Open Redirect payload
-        String maliciousRelayStateUrl = "https://hacker-phishing-site.invalid/steal-credentials";
-
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        String response = new SpidResponseBuilder(mockIdpSpid.XML_RESPONSE_TEMPLATE, spidRequest.getRequestId())
-            .withIdpConfig(identityProvider.signingIdpSsoUrl)
-            .withEntityIds(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT, identityProvider.signingIdpEntityId)
-            .withCertificates(mockIdpSpid.IDP_MOCK_PRIVATE_KEY, mockIdpSpid.IDP_MOCK_CERTIFICATE)
-            .withSignature()
-            .buildResponse();
-
-        // 2. Attacker submits a legitimate SAML Response but injects the malicious URL into the RelayState,
-        // hoping the Service Provider will blindly redirect the user to it upon success.
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", maliciousRelayStateUrl) // Injected malicious URL
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            )
-            .andExpect(status().is3xxRedirection())
-            // 3. Spring Security completely ignores the malicious URL. Instead of redirecting to it,
-            // the flow breaks gracefully because the RelayState does not match the internal session context.
-            // The user is safely redirected back to the local login/error page.
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-    }
-
-    @Test
-    @DisplayName("Verifica runtime HTML Form (HTTP-POST)")
-    public void testRuntimeHtmlFormStructurePost() throws Exception {
-        // 1. Create an active session pre-populated with a protected resource request
-        // This ensures Spring Security will automatically generate a valid RelayState
-        MockHttpSession session = userUtils.createSessionWithSavedClientRequest(BASE_URL);
-
-        // 2. Execute the SSO initialization request directly and capture the raw HTML response
-        String htmlResponse = mockMvc.perform(post(BASE_URL + AUTHENTICATE_PATH + identityProvider.registrationIdPost)
-                .secure(true)
-                .session(session))
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-        // 3. Verify the HTML response contains a form intended for the IdP via POST
-        assertThat(htmlResponse).contains("<form");
-        assertThat(htmlResponse).containsIgnoringCase("method=\"post\"");
-
-        String action = spidProviderConfigRepository.findByProviderId(identityProvider.signingIdpProvider)
-            .getRelyingPartyRegistration(mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_POST)
-            .getAssertingPartyDetails()
-            .getSingleSignOnServiceLocation();
-
-        assertThat(htmlResponse).contains("action=\"" + action + "\"");
-
-        // 4. Verify the presence of the mandatory hidden SAML inputs
-        assertThat(htmlResponse).contains("name=\"SAMLRequest\"");
-        assertThat(htmlResponse).contains("name=\"RelayState\"");
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the SAML Response is unsolicited
-     * (missing the InResponseTo attribute that links it to a specific AuthnRequest).
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Unsolicited Response (InResponseTo assente)")
-    public void testAuthenticationFailsOnUnsolicitedResponse() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Generates a payload without InResponseTo
-        String response = hackerUtils.prepareForSimulationUnsolicitedResponse(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the SAML Response Destination
-     * does not match the actual ACS endpoint of the Service Provider.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Recipient Mismatch (Destination errato)")
-    public void testAuthenticationFailsOnRecipientMismatch() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Generates a payload with a fictitious Destination (Recipient Mismatch)
-        String response = hackerUtils.prepareForSimulationRecipientMismatch(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the SAML Response Issuer
-     * does not match any registered Identity Provider entity ID.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per Issuer Mismatch (EntityID IdP non registrato)")
-    public void testAuthenticationFailsOnIssuerMismatch() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Generates a signed payload but originating from an unknown Issuer
-        String response = hackerUtils.prepareForSimulationIssuerMismatch(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-    }
-
-    /**
-     * Security Test: Ensures authentication fails if the Assertion Conditions
-     * specify a NotBefore timestamp that is in the future.
-     */
-    @Test
-    @DisplayName("Sicurezza: Fallimento atteso per NotBefore futuro (Assertion non ancora valida)")
-    public void testAuthenticationFailsOnNotBeforeFuture() throws Exception {
-        SpidRequest spidRequest = new SpidRequestFlow(mockMvc)
-            .withEndpoints(BASE_URL, USER_DESTINATION_URL, AUTHENTICATE_PATH)
-            .withIdpConfig(identityProvider.registrationIdRedirect)
-            .withSession()
-            .executeRequest();
-
-        // Generates a response where the Assertion's validity begins in the future (e.g., tomorrow)
-        String response = hackerUtils.prepareForSimulationNotValidNotBeforeFuture(
-            spidRequest,
-            mockIdpSpid.XML_RESPONSE_TEMPLATE,
-            identityProvider.signingIdpSsoUrl,
-            mockIdpSpid.ASSERTING_PARTY_ENTITY_ID_REDIRECT,
-            identityProvider.signingIdpEntityId,
-            mockIdpSpid.IDP_MOCK_PRIVATE_KEY,
-            mockIdpSpid.IDP_MOCK_CERTIFICATE
-        );
-
-        this.mockMvc.perform(post(identityProvider.signingIdpSsoUrl)
-                .secure(true)
-                .param("SAMLResponse", response)
-                .param("RelayState", spidRequest.getRelayState())
-                .session(spidRequest.getSession())
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl(LOGIN_DESTINATION_URL));
-
-        // Verifies that Spring Security has raised a time-related authentication exception
-        Exception sessionException = (Exception) spidRequest.getSession().getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
-        assertThat(sessionException).isNotNull();
-        assertThat(sessionException.getMessage()).containsIgnoringCase("is not yet valid");
-    }
-}
